Include prestress stiffness
Calculate spoke stiffness matrix including a contribution from the
initial spoke tension which acts orthogonal to the spoke direction.

This effectively makes the problem non-linear, since the stiffness
depends on the tension, and the tension depends on the stiffness
(through the deformation). However, since the axial stiffness (which
determines the spoke tension) and the tension stiffness (which is
proportional to tension) are orthogonal, only 2 iterations are required
to converge.