Solve generalized eigenvalue problem without scipy.
Most generally, finding the buckling tension reduces to the
generalized eigenvalue problem (GEP):

  (K_matl + T*K_geom)*x = 0

or

  K_matl*x = -T*K_geom*x

K_matl and K_geom are both real symmetric matrices, but neither
are necessarily invertible on their own. However, there exists a
tension t such that K_matl + t*K_geom IS invertible. The
eigenvalues of the original GEP can be solved by solving the
standard eigenvalue problem

  (K_matl + t*K_geom)*x = l_i * x

and then obtaining the eigenvalues of the original GEP by

  T_i = t + 1/l_i

This can be done efficiently with NumPy alone, thus eliminating the
need for an additional package.