Scale strain energy by Rs/Rc
The strain-displacement relation depends on the local radius of curvature.
Before, R_s was used as the approximation for the constant curvature
through the depth of the beam. R_c is a better approximation (geometric
centroid of the cross-section). The resulting energy equation changes by
a factor of (Rs/Rc)^2.

Furthermore, the energy equation needs to be augmented by a factor of
Rc/Rs to account for the fact that the centroid line is longer than the
shear-center line by a factor of Rc/Rs.

The total correction factor is therefore Rs/Rc.